+++ Tesis/qr_flutter/lib/src/registro_cirujias.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:flutter_time_picker_spinner/flutter_time_picker_spinner.dart';
-import 'package:qr_flutter/src/validacionesRegistro.dart';
+import 'package:qr_flutter/validations/validacionesRegistro.dart';
 
 class RegisterPage extends StatefulWidget {
   @override
